BACKEND/ms-auth/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
pydantic==2.10.6
pydantic-settings==2.7.0
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.7